Cambios antes del deploy

+++ index.js
@@ -8,8 +8,14 @@ app.use(express.static(__dirname + "/"));
 
 app.get("/agregarUsuario/:nick", function(request, response) {
     let nick = request.params.nick;
-    let res = sistema.agregarUsuario(nick);
-    response.send(res);
+    let existed = sistema.usuarioActivo(nick);
+    if (existed) {
+        response.status(409).json({ ok: false });
+    } else {
+        let res = sistema.agregarUsuario(nick);
+        response.status(200).json(res);
+    }
+    
 });
 
 app.get("/obtenerUsuarios", function(request, response) {
@@ -30,8 +36,13 @@ app.get("/numeroUsuarios", function(request, response) {
 
 app.get("/eliminarUsuario/:nick", function(request, response) {
     let nick = request.params.nick;
-    sistema.eliminarUsuario(nick);
-    response.send("Usuario eliminado");
+    let existed = sistema.usuarioActivo(nick);
+    if (existed) {
+        sistema.eliminarUsuario(nick);
+        response.json({ ok: true });
+    } else {
+        response.json({ ok: false });
+    }
 });
 
 app.listen(PORT, () => {
